refactor: item styles
